refactor: reorganize DICOM utilities and input module
