+++ src/imgtools/dicom/__init__.py
@@ -1,11 +1,9 @@
 from .dicom_find import find_dicoms
-from .input import (
-    extract_roi_meta,
-    extract_roi_names,
+from .dicom_reader import (
+    DicomInput,
     load_dicom,
     load_rtstruct_dcm,
     load_seg_dcm,
-    rtstruct_reference_uids,
 )
 from .utils import lookup_tag, similar_tags, tag_exists
 
@@ -15,11 +13,9 @@ __all__ = [
     "lookup_tag",
     "similar_tags",
     "tag_exists",
-    # input
+    # dicom_reader
+    "DicomInput",
     "load_dicom",
     "load_rtstruct_dcm",
     "load_seg_dcm",
-    "extract_roi_meta",
-    "extract_roi_names",
-    "rtstruct_reference_uids",
 ]
