+++ src/imgtools/dicom/modality_utils/__init__.py
@@ -1,9 +1,3 @@
-from .dicom_reader import (
-    load_dicom,
-    load_rtstruct_dcm,
-    load_seg_dcm,
-    path_from_pathlike,
-)
 from .rtdose_utils import (
     RTDOSERefPlanSOP,
     RTDOSERefSeries,
@@ -31,10 +25,6 @@ from .sr_utils import (
 )
 
 __all__ = [
-    "load_dicom",
-    "path_from_pathlike",
-    "load_rtstruct_dcm",
-    "load_seg_dcm",
     # rtstruct_utils
     "extract_roi_meta",
     "extract_roi_names",
