+++ src/imgtools/dicom/modality_utils/rtplan_utils.py
@@ -1,4 +1,4 @@
-from imgtools.dicom.input.dicom_reader import DicomInput, load_dicom
+from imgtools.dicom.dicom_reader import DicomInput, load_dicom
 from imgtools.logging import logger
 
 __all__ = ["rtplan_reference_uids", "RTPLANRefStructSOP"]
