+++ src/imgtools/dicom/modality_utils/sr_utils.py
@@ -1,4 +1,4 @@
-from imgtools.dicom.input.dicom_reader import DicomInput, load_dicom
+from imgtools.dicom.dicom_reader import DicomInput, load_dicom
 
 """
 StructuredReport (SR)
