+++ src/imgtools/dicom/read_tags.py
@@ -5,20 +5,11 @@ and reading specific tags from DICOM files.
 
 Functions
 ---------
-truncate_uid(uid: str, last_digits: int = 5) -> str
-    Truncate a UID to the last n characters.
 read_tags(file: Path, tags: list[str], truncate: bool = True) -> dict[str, str]
     Read specified tags from a DICOM file.
 
 Examples
 --------
-Truncate a UID:
-    >>> truncate_uid(
-    ...     "1.2.840.10008.1.2.1",
-    ...     last_digits=5,
-    ... )
-    '.1.2.1'
-
 Read tags from a DICOM file:
     >>> from pathlib import (
     ...     Path,
@@ -37,48 +28,8 @@ Read tags from a DICOM file:
 from pathlib import Path
 from typing import Dict, List, Optional
 
-from pydicom import dcmread
-from pydicom.errors import InvalidDicomError
-
-
-def truncate_uid(uid: str, last_digits: int = 5) -> str:
-    """
-    Truncate the UID to the last n characters (including periods and underscores).
-
-    If the UID is shorter than `last_digits`, the entire UID is returned.
-
-    Parameters
-    ----------
-    uid : str
-        The UID string to truncate.
-    last_digits : int, optional
-        The number of characters to keep at the end of the UID (default is 5).
-
-    Returns
-    -------
-    str
-        The truncated UID string.
-
-    Examples
-    --------
-    >>> truncate_uid(
-    ...     "1.2.840.10008.1.2.1",
-    ...     last_digits=5,
-    ... )
-    '.1.2.1'
-    >>> truncate_uid(
-    ...     "12345",
-    ...     last_digits=10,
-    ... )
-    '12345'
-    """
-    assert uid is not None
-    assert isinstance(uid, str)
-    assert isinstance(last_digits, int)
-    if last_digits >= len(uid) or last_digits <= 0:
-        return uid
-
-    return uid[-last_digits:]
+from imgtools.dicom import load_dicom
+from imgtools.utils import truncate_uid
 
 
 def read_tags(
@@ -168,19 +119,9 @@ def read_tags(
         and tags is not None
     )
 
-    try:
-        dicom = dcmread(
-            file, specific_tags=tags, stop_before_pixels=True, force=force
-        )
-    except FileNotFoundError as fnfe:
-        errmsg = f"File not found: {file}"
-        raise FileNotFoundError(errmsg) from fnfe
-    except InvalidDicomError as ide:
-        errmsg = f"Invalid DICOM file: {file}"
-        raise InvalidDicomError(errmsg) from ide
-    except ValueError as ve:
-        errmsg = f"Value error reading DICOM file: {file}"
-        raise ValueError(errmsg) from ve
+    dicom = load_dicom(
+        file, force=force, stop_before_pixels=True, specific_tags=tags
+    )
 
     result = {}
