--- src/imgtools/io/loaders/__init__.py
@@ -1,31 +1,0 @@
-from .old_loaders import (
-    BaseLoader,
-    ImageCSVLoader,
-    ImageFileLoader,
-    ImageTreeLoader,
-    auto_dicom_result,
-    read_dicom_auto,
-    read_dicom_pet,
-    read_dicom_rtdose,
-    read_dicom_rtstruct,
-    read_dicom_scan,
-    read_dicom_seg,
-    read_dicom_series,
-    read_image,
-)
-
-__all__ = [
-    "read_image",
-    "read_dicom_series",
-    "read_dicom_scan",
-    "read_dicom_rtstruct",
-    "read_dicom_rtdose",
-    "read_dicom_pet",
-    "read_dicom_seg",
-    "read_dicom_auto",
-    "auto_dicom_result",
-    "BaseLoader",
-    "ImageTreeLoader",
-    "ImageCSVLoader",
-    "ImageFileLoader",
-]
